#resolve #123
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1731062 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 973b23b0644c715f335d4997e10120c8499d3462

+++ src/jorphan/org/apache/jorphan/util/JOrphanUtils.java
@@ -570,9 +570,10 @@ public final class JOrphanUtils {
 
     /**
      * @param elapsedSec long elapsed time in seconds
-     * @return String formated with format HH:mm:ss
+     * @return String formatted with format HH:mm:ss
      */
     public static String formatDuration(long elapsedSec) {
-        return String.format("%02d:%02d:%02d", Long.valueOf(elapsedSec / 3600), Long.valueOf((elapsedSec % 3600) / 60), Long.valueOf(elapsedSec % 60));
+        return String.format("%02d:%02d:%02d",
+                elapsedSec / 3600, (elapsedSec % 3600) / 60, elapsedSec % 60);
     }
 }
